download function move to utils, and update readme
